api/requirements.txt
Flask==2.1.2
flask-cors==3.0.10
transformers==4.30.2
Pillow==9.2.0
torch==2.0.1
